Relocate util functions
