+++ src/types.ts
@@ -19,7 +19,6 @@ import {
   BrowseRequestType,
   FacetOption as ApiFacetOption,
 } from '@constructor-io/constructorio-client-javascript/lib/types';
-import { MakeOptional } from './utils/typeHelpers';
 
 export {
   Nullable,
@@ -31,6 +30,12 @@ export {
   ApiFacetOption,
   ApiGroup,
 };
+
+/**
+ * Given a Type T and a set of keys K (pipe-delimited string), make those keys optional.
+ */
+export type MakeOptional<Type, Keys extends string & keyof Partial<Type>> = Omit<Type, Keys> &
+  Partial<Pick<Type, Keys>>;
 
 export interface ApiHierarchicalFacetOption extends ApiFacetOption {
   options: Array<ApiHierarchicalFacetOption>;
