+++ src/utils/index.ts
@@ -4,5 +4,5 @@ export * from './styleHelpers';
 export * from './transformers';
 export * from './typeHelpers';
 export * from './urlHelpers';
-export * from './pageHelpers';
-export * from './utilFunctions';
+export * from './plpRequestHelpers';
+export * from './common';
